requirements.txt
# LLM & AI
# google-genai is specified in pyproject.toml

# Web Framework
streamlit>=1.31.0

# Database
pymongo>=4.6.0

# Data Processing
pandas>=2.2.0
python-dateutil>=2.8.2

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
requests>=2.31.0
